o_o , un cast de long int en int non indiqué par le compilateur peut provoquer des bugs incompréhensibles. Surtout sur un AVR 8-bit........

+++ baseRoulante/Arduino/CarteAsservissementListeConsignes/Manager.h
@@ -73,8 +73,8 @@ class Manager {
 		long int	distanceBkp;
 
 		typedef struct {
-			int distance;
-			int angle;
+			long int distance;
+			long int angle;
 		}Consigne;
 
 		typedef struct {
